GET /api/v1/users/<int:pk>, POST /api/v1/users

+++ tweets/views.py
@@ -4,9 +4,12 @@ from .serializer import TweetSerializer
 from rest_framework.views import APIView
 from rest_framework.exceptions import NotFound
 from rest_framework.status import HTTP_204_NO_CONTENT
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 
 
 class Tweets(APIView):
+
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get(self, request):
         tweets = Tweet.objects.all()
@@ -24,6 +27,8 @@ class Tweets(APIView):
 
 class TweetDetail(APIView):
 
+    permission_classes = [IsAuthenticatedOrReadOnly]
+
     def get_object(self, tweet_id):
         try:
             return Tweet.objects.get(pk=tweet_id)
@@ -37,6 +42,8 @@ class TweetDetail(APIView):
 
     def put(self, request, tweet_id):
         tweet = self.get_object(tweet_id)
+        if tweet.user != request.user:
+            raise PermissionError
         serializer = TweetSerializer(tweet, data=request.data)
         if serializer.is_valid():
             updated_tweet = serializer.save()
@@ -46,5 +53,7 @@ class TweetDetail(APIView):
 
     def delete(self, request, tweet_id):
         tweet = self.get_object(tweet_id)
+        if tweet.user != request.user:
+            raise PermissionError
         tweet.delete()
         return Response(HTTP_204_NO_CONTENT)
